Fix bug where empty options doesn't parse default
Cases where there is no input arguments and options causes the
default to be skipped. This is because the input string is empty
which causes a short circuit in the code. This has been fixed

NOTE: Global options are parsed seperately and are removed by this
point regardless what else is entered.