Fix bug that converts decimal to octal
Previously if an option had a type of OptionParser::DecimalInteger
it would still convert as an ocatal if the string values starts with
a '0'. E.g. '035' would still convert to 29 even if the type was
DecimalInteger.

This commit patches OptionParser to force the DecimalIntegers to be
converted in base 10 regardless what the string is.