Revert version numbering to match forked project
To prevent confusion in version numbering, the version has been reverted
to what it was originally. At the time this project was forked, the
version number was 4.4.3.

To distinguish between the original project and our updates, an
additional "alces" flag has been added.

For this reason the version number is "4.4.3-alces0". This way
the alces numbering can be bumped in the future.

NOTE: The previous version numbering will be maintained as older
versions of metalware depend on it.

https://github.com/alces-software/commander/pull/2#discussion_r168198886