Turn the ValidateInputs patch off for vanilla test
The original tests are incompatible with the ValidateInputs patches. The
easiest way to get them passing again is to turn the patch off when they
are running.

This is done by adding a new `PatchEnabled` constant that is set to
true. This constant is then stubbed in the original tests to be false.