Fix bug where sub commands help can not be displayed
The integration of checking the inputs has broken the sub commands. This
is because `Commander` reexec a `commander help` command to display the
help pages.

However this internal `Commander` command does not pass the syntax
checking and thus errors. The fix is to not run the validation on this
internal `command`.

Hopefully this works, Commander is confusing AF